Move RpcClient related code from staking/ to watcher/
prepare to merge crosschain branch into main
