+++ cmd/staking/showopreturn/main.go
@@ -2,12 +2,13 @@ package main
 
 import (
 	"fmt"
-	"github.com/tendermint/tendermint/libs/log"
 	"os"
 	"strconv"
 	"strings"
 
-	"github.com/smartbch/smartbch/staking"
+	"github.com/tendermint/tendermint/libs/log"
+
+	"github.com/smartbch/smartbch/watcher"
 )
 
 // To summmarize all the op_return vouts since the first SLP transaction:
@@ -36,12 +37,12 @@ func main() {
 		panic(err)
 	}
 
-	client := staking.NewRpcClient(rpcURL, rpcUsername, rpcPassword, "text/plain;", log.NewNopLogger())
+	client := watcher.NewRpcClient(rpcURL, rpcUsername, rpcPassword, "text/plain;", log.NewNopLogger())
 	printAllOpReturn(client, startH, endH)
 	//client.PrintAllOpReturn(519995, 679995)
 }
 
-func printAllOpReturn(client *staking.RpcClient, startHeight, endHeight int64) {
+func printAllOpReturn(client *watcher.RpcClient, startHeight, endHeight int64) {
 	for h := startHeight; h < endHeight; h++ {
 		fmt.Printf("Height: %d\n", h)
 		hash, err := client.GetBlockHash(h)
