+++ watcher/rpcclient.go
@@ -1,17 +1,18 @@
-package staking
+package watcher
 
 import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
-	"github.com/tendermint/tendermint/libs/log"
 	"io/ioutil"
 	"net/http"
 	"strings"
 
 	"github.com/ethereum/go-ethereum/common/hexutil"
+	"github.com/tendermint/tendermint/libs/log"
 
-	"github.com/smartbch/smartbch/staking/types"
+	stakingtypes "github.com/smartbch/smartbch/staking/types"
+	"github.com/smartbch/smartbch/watcher/types"
 )
 
 const (
@@ -20,112 +21,7 @@ const (
 	ReqStrBlock      = `{"jsonrpc": "1.0", "id":"smartbch", "method": "getblock", "params": ["%s"] }`
 	ReqStrTx         = `{"jsonrpc": "1.0", "id":"smartbch", "method": "getrawtransaction", "params": ["%s", true, "%s"] }`
 	ReqStrEpochs     = `{"jsonrpc": "2.0", "method": "sbch_getEpochs", "params": ["%s","%s"], "id":1}`
-	Identifier       = "73424348" // ascii code for 'sBCH'
-	Version          = "00"
 )
-
-type JsonRpcError struct {
-	Code    int    `json:"code"`
-	Message string `json:"message"`
-}
-
-type BlockCountResp struct {
-	Result int64         `json:"result"`
-	Error  *JsonRpcError `json:"error"`
-	Id     string        `json:"id"`
-}
-
-type BlockHashResp struct {
-	Result string        `json:"result"`
-	Error  *JsonRpcError `json:"error"`
-	Id     string        `json:"id"`
-}
-
-type BlockInfo struct {
-	Hash              string   `json:"hash"`
-	Confirmations     int      `json:"confirmations"`
-	Size              int      `json:"size"`
-	Height            int64    `json:"height"`
-	Version           int      `json:"version"`
-	VersionHex        string   `json:"versionHex"`
-	Merkleroot        string   `json:"merkleroot"`
-	Tx                []string `json:"tx"`
-	Time              int64    `json:"time"`
-	MedianTime        int64    `json:"mediantime"`
-	Nonce             int      `json:"nonce"`
-	Bits              string   `json:"bits"`
-	Difficulty        float64  `json:"difficulty"`
-	Chainwork         string   `json:"chainwork"`
-	NumTx             int      `json:"nTx"`
-	PreviousBlockhash string   `json:"previousblockhash"`
-}
-
-type BlockInfoResp struct {
-	Result BlockInfo     `json:"result"`
-	Error  *JsonRpcError `json:"error"`
-	Id     string        `json:"id"`
-}
-
-type CoinbaseVin struct {
-	Coinbase string `json:"coinbase"`
-	Sequence int    `json:"sequence"`
-}
-
-type Vout struct {
-	Value        float64                `json:"value"`
-	N            int                    `json:"n"`
-	ScriptPubKey map[string]interface{} `json:"scriptPubKey"`
-}
-
-type TxInfo struct {
-	TxID          string                   `json:"txid"`
-	Hash          string                   `json:"hash"`
-	Version       int                      `json:"version"`
-	Size          int                      `json:"size"`
-	Locktime      int                      `json:"locktime"`
-	VinList       []map[string]interface{} `json:"vin"`
-	VoutList      []Vout                   `json:"vout"`
-	Hex           string                   `json:"hex"`
-	Blockhash     string                   `json:"blockhash"`
-	Confirmations int                      `json:"confirmations"`
-	Time          int64                    `json:"time"`
-	BlockTime     int64                    `json:"blocktime"`
-}
-
-func (ti TxInfo) GetValidatorPubKey() (pubKey [32]byte, success bool) {
-	for _, vout := range ti.VoutList {
-		asm, ok := vout.ScriptPubKey["asm"]
-		if !ok || asm == nil {
-			continue
-		}
-		script, ok := asm.(string)
-		if !ok {
-			continue
-		}
-		prefix := "OP_RETURN " + Identifier + Version
-		if !strings.HasPrefix(script, prefix) {
-			continue
-		}
-		script = script[len(prefix):]
-		if len(script) != 64 {
-			continue
-		}
-		bz, err := hex.DecodeString(script)
-		if err != nil {
-			continue
-		}
-		copy(pubKey[:], bz)
-		success = true
-		break
-	}
-	return
-}
-
-type TxInfoResp struct {
-	Result TxInfo        `json:"result"`
-	Error  *JsonRpcError `json:"error"`
-	Id     string        `json:"id"`
-}
 
 type RpcClient struct {
 	url         string
@@ -148,42 +44,12 @@ func NewRpcClient(url, user, password, contentType string, logger log.Logger) *R
 	}
 }
 
-func (client *RpcClient) sendRequest(reqStr string) ([]byte, error) {
-	body := strings.NewReader(reqStr)
-	req, err := http.NewRequest("POST", client.url, body)
-	if err != nil {
-		return nil, err
-	}
-	req.SetBasicAuth(client.user, client.password)
-	req.Header.Set("Content-Type", client.contentType)
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	respData, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-	resp.Body.Close()
-	return respData, nil
-}
-
 func (client *RpcClient) GetLatestHeight() (height int64) {
 	height, client.err = client.getCurrHeight()
 	if client.err != nil {
 		client.logger.Debug("GetLatestHeight failed", client.err.Error())
 	}
 	return
-}
-
-func (client *RpcClient) GetBlockHash(height int64) (string, error) {
-	return client.getBlockHashOfHeight(height)
-}
-func (client *RpcClient) GetBlockInfo(hash string) (*BlockInfo, error) {
-	return client.getBlock(hash)
-}
-func (client *RpcClient) GetTxInfo(hash string, blockhash string) (*TxInfo, error) {
-	return client.getTx(hash, blockhash)
 }
 
 func (client *RpcClient) GetBlockByHeight(height int64) *types.BCHBlock {
@@ -208,7 +74,7 @@ func (client *RpcClient) GetBlockByHash(hash [32]byte) *types.BCHBlock {
 	return blk
 }
 
-func (client *RpcClient) GetEpochs(start, end uint64) []*types.Epoch {
+func (client *RpcClient) GetEpochs(start, end uint64) []*stakingtypes.Epoch {
 	epochs := client.getEpochs(start, end)
 	if client.err != nil {
 		client.logger.Debug("GetEpochs failed", client.err.Error())
@@ -216,8 +82,28 @@ func (client *RpcClient) GetEpochs(start, end uint64) []*types.Epoch {
 	return epochs
 }
 
+func (client *RpcClient) sendRequest(reqStr string) ([]byte, error) {
+	body := strings.NewReader(reqStr)
+	req, err := http.NewRequest("POST", client.url, body)
+	if err != nil {
+		return nil, err
+	}
+	req.SetBasicAuth(client.user, client.password)
+	req.Header.Set("Content-Type", client.contentType)
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		return nil, err
+	}
+	respData, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+	resp.Body.Close()
+	return respData, nil
+}
+
 func (client *RpcClient) getBCHBlock(hash string) *types.BCHBlock {
-	var bi *BlockInfo
+	var bi *types.BlockInfo
 	bi, client.err = client.getBlock(hash)
 	if client.err != nil {
 		return nil
@@ -238,14 +124,14 @@ func (client *RpcClient) getBCHBlock(hash string) *types.BCHBlock {
 		return nil
 	}
 	if bi.Height > 0 {
-		var coinbase *TxInfo
+		var coinbase *types.TxInfo
 		coinbase, client.err = client.getTx(bi.Tx[0], bi.Hash)
 		if client.err != nil {
 			return nil
 		}
 		pubKey, ok := coinbase.GetValidatorPubKey()
 		if ok {
-			nomination := types.Nomination{
+			nomination := stakingtypes.Nomination{
 				Pubkey:         pubKey,
 				NominatedCount: 1,
 			}
@@ -260,7 +146,7 @@ func (client *RpcClient) getCurrHeight() (int64, error) {
 	if err != nil {
 		return -1, err
 	}
-	var blockCountResp BlockCountResp
+	var blockCountResp types.BlockCountResp
 	err = json.Unmarshal(respData, &blockCountResp)
 	if err != nil {
 		return -1, err
@@ -277,7 +163,7 @@ func (client *RpcClient) getBlockHashOfHeight(height int64) (string, error) {
 	if err != nil {
 		return "", err
 	}
-	var blockHashResp BlockHashResp
+	var blockHashResp types.BlockHashResp
 	err = json.Unmarshal(respData, &blockHashResp)
 	if err != nil {
 		return "", err
@@ -289,12 +175,12 @@ func (client *RpcClient) getBlockHashOfHeight(height int64) (string, error) {
 	return blockHashResp.Result, nil
 }
 
-func (client *RpcClient) getBlock(hash string) (*BlockInfo, error) {
+func (client *RpcClient) getBlock(hash string) (*types.BlockInfo, error) {
 	respData, err := client.sendRequest(fmt.Sprintf(ReqStrBlock, hash))
 	if err != nil {
 		return nil, err
 	}
-	var blockInfoResp BlockInfoResp
+	var blockInfoResp types.BlockInfoResp
 	err = json.Unmarshal(respData, &blockInfoResp)
 	if err != nil {
 		return nil, err
@@ -306,12 +192,12 @@ func (client *RpcClient) getBlock(hash string) (*BlockInfo, error) {
 	return &blockInfoResp.Result, nil
 }
 
-func (client *RpcClient) getTx(hash string, blockhash string) (*TxInfo, error) {
+func (client *RpcClient) getTx(hash string, blockhash string) (*types.TxInfo, error) {
 	respData, err := client.sendRequest(fmt.Sprintf(ReqStrTx, hash, blockhash))
 	if err != nil {
 		return nil, err
 	}
-	var txInfoResp TxInfoResp
+	var txInfoResp types.TxInfoResp
 	err = json.Unmarshal(respData, &txInfoResp)
 	if err != nil {
 		return nil, err
@@ -338,7 +224,7 @@ type smartBchJsonrpcMessage struct {
 	Result  json.RawMessage       `json:"result,omitempty"`
 }
 
-func (client *RpcClient) getEpochs(start, end uint64) []*types.Epoch {
+func (client *RpcClient) getEpochs(start, end uint64) []*stakingtypes.Epoch {
 	var respData []byte
 	respData, client.err = client.sendRequest(fmt.Sprintf(ReqStrEpochs, hexutil.Uint64(start).String(), hexutil.Uint64(end).String()))
 	if client.err != nil {
@@ -349,10 +235,20 @@ func (client *RpcClient) getEpochs(start, end uint64) []*types.Epoch {
 	if client.err != nil {
 		return nil
 	}
-	var epochsResp []*types.Epoch
+	var epochsResp []*stakingtypes.Epoch
 	client.err = json.Unmarshal(m.Result, &epochsResp)
 	if client.err != nil {
 		return nil
 	}
 	return epochsResp
+}
+
+func (client *RpcClient) GetBlockHash(height int64) (string, error) {
+	return client.getBlockHashOfHeight(height)
+}
+func (client *RpcClient) GetBlockInfo(hash string) (*types.BlockInfo, error) {
+	return client.getBlock(hash)
+}
+func (client *RpcClient) GetTxInfo(hash string, blockhash string) (*types.TxInfo, error) {
+	return client.getTx(hash, blockhash)
 }
